+++ ros2plot/ros/__init__.py
@@ -1,4 +1,8 @@
 
 from .multisub import MultiSubscriber
+from .message_handlers import get_message_handler, get_message_processor, get_message_2d_plottable_fields
 
-__all__ = ["MultiSubscriber"]
+__all__ = [ "MultiSubscriber",
+            "get_message_handler",
+            "get_message_processor",
+            "get_message_2d_plottable_fields"]